chore: change import path to fix unit tests
