+++ packages/core/client/src/modules/blocks/data-blocks/details-single/detailsBlockSettings.ts
@@ -9,13 +9,10 @@
 
 import { useFieldSchema } from '@formily/react';
 import { SchemaSettings } from '../../../../application/schema-settings/SchemaSettings';
-import { useCollection_deprecated } from '../../../../collection-manager';
-import {
-  SchemaSettingsBlockTitleItem,
-  SchemaSettingsFormItemTemplate,
-  SchemaSettingsLinkageRules,
-} from '../../../../schema-settings';
 import { SchemaSettingsItemType } from '../../../../application/schema-settings/types';
+import { useCollection_deprecated } from '../../../../collection-manager';
+import { SchemaSettingsFormItemTemplate, SchemaSettingsLinkageRules } from '../../../../schema-settings';
+import { SchemaSettingsBlockTitleItem } from '../../../../schema-settings/SchemaSettingsBlockTitleItem';
 
 const commonItems: SchemaSettingsItemType[] = [
   {
